refactor: pluralized helper folder name
